EmbeddedQueryDependencyLinksStore to create object ID if not available, refs 1117
For example if a new query is created with entities that have not yet been
used before (== new) then the SQLStore does not have nor will it create
an ID for something that is not an annotation (a query description is not
an annotation).

While this is an edge case (where a query is created before any
annotation is available) nevertheless we start to track its dependency
early.