Exported API time format is insufficient
The simple `getMwTimestamp` output (e.g. `-18208281600` for `1 Jan 1393`)
is insufficient for a general workable API solution in order for a
consumer to create a meaningful interpretation.

The output was extended to carry two representations, the previous
timestamp and now the raw format (e.g. `2/1393/1/1`) which can can be
used to deserialize the data into its original value depiction
(with the intended precision).

`smw.dataItem.time` as been adjusted to find the correct
timestamp element.

Consumers that use JS/`smw.dataItem.time` can work with both
API results (< 0.7 >).

```
"Has date": [
    {
        "timestamp": "-2000101000000",
        "raw": "2/-200"
    },
    {
        "timestamp": "-18208281600",
        "raw": "2/1393/1/1"
    }
```