Add extra search column to support case-insensitive regex searches
Instead of using the sortkey to match a text string, any
regex search (if enabled) will use the normalized search index to allow
for case-insensitive result matches.

A query condition `[[Has text::~lorem*]]` will return the same results
as for a `[[Has text::~LOREM*]]` (or `[[Has text::~lOrEm*]]`)
condition.

It is required that:

- `$GLOBALS['smwgEnabledCaseInsensitiveRegExMatchSearch']` is set true
- `update.php` and `rebuildData.php` both run after the setting was enabled

Test cases include:

- `[[Has text::~lorem*]]` and `[[Has text::~*REM*]]`
- `[[Has text::!~*REM*]]`
- `[[Has url::~*rem*]]` and `[[Has url::~*REm*]]`
- `[[~*/6/lore*]]` and `[[~*/6/lORe*]]`
- `[[Has text::~*æå*]]`