PropertyValue to be fowarded to its redirect target
A property `Has name` is being redirected to `Foaf:name` which
makes the following query not to display any pr-values (because `Has name`
does not hold any value reference) therefore if a redirect target is known
use it instead.

```
{{#ask: [[Has name::John Doe]]
 |?Has name
}}
```

A `PropertyValue` will after this PR always point to the end target.

refs 1244