Support custom record types
This is not an ideal solution (cannot validate for an instance at the
time where the check occurs) however it allows extensions that define a
record subtype and still use the same functionality.

The convention is that additional record types to be registered
with `_rec` being a part of the datatype ID (__foo_rec etc).