Add possibility to specify precision for a numeric datatype, refs 1125
`[[Display precision of::<int>]]` can declare a scale for decimal precision
to a numeric property type so that display output can be restricted to
something like 2 digits as in `€ 42.45, € 17.14, € 60.00`.

`Display precision of` only restricts output representation and does not
interfere with the internal "full" precision.

It is expected that values are transformed using:

`[[Display precision of::2]]`

- 0.001 -> 0.00
- 0.005 -> 0.01
- € 60  -> € 60.00
- 30.445 € -> € 30.45 (3.00 ¥, 32.27 $, 21.31 £)
- 100 °F -> 310.93 K (37.78 °C, 100.00 °F, 559.67 °R)

`[[Display precision of::4]]`

- 100 °F -> 310.9278 K (37.7778 °C, 100.0000 °F, 559.6700 °R)
- 30.445 € -> € 30.4450 (3.0000 ¥, 32.2717 $, 21.3115 £)