Add localizable context help for properties, refs 1344
"Has property description" is a predefined property and can be used to
describe a local property in context of a language.

If a user switches its language preference (`wgLang`) and a property with
description matches the language then it will be used during
the highlight display.

If a predefined property contains a local "Has property description" then
this information will be displayed instead of the standard message.

```
Property: SomeProperty

[[Has property description::Is a property that ...@en]]
[[Has property description::このプロパティは。。。@ja]]

```
The `browsebyproperty` API has been amended to contain description information
such as:

```
api.php?action=browsebyproperty&property=SomeProperty&lang=en

{
    "query": {
        "Has_date": {
            "label": "SomeProperty",
            "key": "SomeProperty",
            "isUserDefined": "",
            "usageCount": 5343,
            "description": {
                "en": "Is a property that ..."
            }
        }
    },
    "version": 0.2,
    "query-continue-offset": 0,
    "meta": {
        "limit": 50,
        "count": 1
    }
}
```