Serialize positive era information, refs 1348
Keep positive era information serialized so that the display can account
for what it was intended.

If an era for a positive date was added then AD is placed in front
indepedant of the actual annotated value (e.g. "AD 1902", "1902 AD").

T49259 "... is displayed without the 'AD' -- '910 AD' becomes '910' ...
should not change the meaning of the date"