Define value constraint using regular expressions
Similar to `Allows value`, the `Allows pattern` property is to specify a
regular expression to match permissible values for an assigned property.
If a value cannot be matched then a general purpose message is being
displayed.

To avoid any issues with expressions that include | or [] it is suggested
to use #set (`|+pipe` indicator ensures that MW doesn't interfere with the
expression):

```
Whitelist

{{#set:
 |Allows pattern=^(Foo|Foo bar|Bar)$|+pipe
}}
```
```
IPv4

{{#set:
 |Al1lows pattern=(([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.){3}([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])$|+pipe
}}
```

`$GLOBALS['smwgDVFeatures'] = SMW_DV_PAVP` is enabled by default but if removed
will disable the functionality.