Allow to track uniqueness of values, refs 178
Sometimes it is necessary to only allow unique values to be assignable to
a property (as for authority references).

For example, an authority reference to an entity like
"Johann Sebastian Bach" (NDL ID 00432003 [0]) should only happen once.

The flag `SMW_DV_PVUC` is required to be added to
`$GLOBALS['smwgDVFeatures']` (is not set by default) to enable the feature
for general use.

A property which is expected to carry a uniqueness trait need to specify
`[[Has uniqueness constraint::true]]`.

The first value the "system" finds and is not yet assigned will be the
established (or unqiue) value with any value of the same literal
representation being compared against.

[0] 国立国会図書館典拠データ検索, http://id.ndl.go.jp/auth/ndlna/00432003