Add search link for zero properties
1216 disposes outdated entities that have no remaining references but it
is not always easy to find some "hidden" references that prevents
the removal when the property has no obvious value annotation.

`PropertyTableIdReferenceFinder` tries to scan related tables to
find at least "one" reference that prevents the
`PropertyTableOutdatedReferenceDisposer` from exercising the removal.

`Special:Properties` is extended to actual display a link
to ease the search.