Add statistics on outdated entities, refs 1100
Outdated entities are not removed immediately from the `ID_TABLE` and in
instead "marked for removal" using the `smw:delete` indicator.

An administrator or curator (SMW) should ensure that this number is
kept low by running the `rebuildData.php` script on a regular basis.

It is also an indicator (besides the error count) whether an
installation is maintained routinately or not.