Preserve spaces in unit labels, refs 1529
Adds the `SMW_DV_NUMV_USPACE` flag for preserving spaces in unit labels
which up to now has been squashed. For example, `sq mi` (official notation)
is converted to `sqmi` without the flag.

In order to allow for different spellings a user is expected to maintain
a list of permitted notations such as `[[Corresponds to::1 km², km ², k m ²]]`
with the first declaration to mark the "standard" output norm.

To enable this feature `smwgDVFeatures` requires the `SMW_DV_NUMV_USPACE`
flag (disabled by default to avoid BC as it is required to list
all norms explicitly).