Add ReferenceValue, refs 1808
- `ReferenceValue` and `ReferenceValueFormatter`
- `smwgDataTypePropertyExemptionList` to exclude listed
`DataType`'s from being registered as property. In case of `Reference` the
use as property to match the `DataType` semantics is of no help since
the definition requires `Has fields` and it is expected that the
`Reference` name to be a common property name hence we refrain from using
it unless in a `DataType` context.