TextByChangeUpdater change update approach, refs 1481 (#2388)
A certain update sequence could allow for a text element to remain in the
index due to the missing information in the diff record about a removed
element.

Instead of using the insert-diff to replace text elements, use the delete-diff
to remove all items from the index and after that insert related data from
the primary data container. This approach ensures the content is in sync with
the SemanticData primary data update.