Construct this.date after deterimining components
Prior to this commit, this.date was constructed with just the year and
then used setXYZ() methods (e.g. setMonth(), setDate()) to update the date
as required based upon other values. This was causing bad dates as noticed
in SemanticResultFormats eventcalendar format, where events showed up
approximately a year in the past. This commit changes it so this.date is
constructed with the full date information rather than being updated one
component at a time.