Update .travis.yml
So they killed old branches in an effort to make digital preservation much easier. Besides this was subject of two talks during the past WikidataCon and this appears imho to be an rather unexpected conclusion to me. However in our context I updated the branches to the ones SMW 3.0 will be supporting, i.e. I added four to the obsolete branches.