Remove SMW\MediaWiki\Job::getTitle()
This function was added in SMW 1.9 because it was missing from
MW 1.19; it was added in MW 1.21 hence it is in all MW versions
supported by SMW 3.0+. It is mandatory to remove this function
for compatibility with MW 1.34 because the method was declared
final. The implementation was always be the same.

Issue: #3914