Allow restoring behavior for format=list
SMW <3.0 wikis that make extensive use of format=list cannot easily
be upgraded due to the breaking changes SMW 3.0 brings in the form
of added HTML in the list output. In big wikis switching over to
plainlist can be an inpractical amount of work (even with search and replace)
and be very error prone.

This change renames list to richlist and adds list as an alias to richlist.
It thus does not change SMW 3.1 behavior. It however does allow the user
to restore pre-3.0 behavior via the $smwgResultAliases setting:

    $smwgResultAliases['richlist'] = [];
    $smwgResultAliases['plainlist'][] = ['list'];