Fix @annotation behavior when used with Page Forms
Manually verified and not sure how to write a non-browser
automated test for this.

Ideally this code would be further refactored so it does
not break whenever it is used together with an extension
that adds an action. Might currently also be broken with
some visual editor stuff, did not check.