CachedFactbox: Avoid master DB connection on read requests
Currently, CachedFactbox fetches the revision ID of the current page via
RevisionGuard, which always uses the master DB, even on read requests. This
results in warnings from MediaWiki's TransactionProfiler ("Expectation
(masterConns <= 0) by MediaWiki::main not met (actual: 1)"), and can pose an
issue if the master DB is in a remote datacenter and so expensive (latency) to
access, or if it is down for maintenance or otherwise unavailable.

Instead, use OutputPage::getRevisionId to fetch the revision ID here, which
avoids this behavior.