Do not rely on MW initialization order when being autoloaded.
Currently SMW setup uses both composer and extension.json.
During the composer autoload it executes SemanticMediaWiki.php
to export it's own constants to be used in LocalSettings.php.
Additionally it does self-check to see if it's intalled correctly.

During autoloading you can't rely on ANY part of MediaWiki to be
initialized. In this case it relied on $wgVersion and $wgLanguageCode
to be present. They will not be anymore since core is moving
dependency autoloading earlier.

In general, this pattern will be prohibited in MW 1.38 and will
be broken intentionally - relying on MW to be partially initialized
during class autoloading is a bad, bad, bad idea.

This fix is a stop-gap to avoid breaking translatewiki entirely.

Bug: https://phabricator.wikimedia.org/T295883