PropertyPage: Avoid primary DB access when viewing property pages
PropertyPage currently fetches semantic data for the page being viewed
via EditInfo, which is more suited for use in an edit context, e.g.
edit-related hooks that need to operate on the result of an edit, as it
always triggers a primary DB connection internally. Instead, fetch the
semantic data directly from the ParserOutput of the property page here,
as we're not in an edit context.