Stabilize tests on MW 1.39 (#5589)
Most of the error spam in the 1.39 test run is ultimately due to
services not being reset between tests, which (a) can cause errors in
itself due to stale data as seen in #5199 and co. and (b) can cause
resulting error states to propagate to other tests. Previous attempts at
fixing this took the approach of selectively resetting services that
were found to be problematic, such as ActorStore. However, this is not
in itself sufficient: not only is the task of identifying services not
safe for reuse between tests quite Sisyphean, resetting them won't reset
services or static singletons that injected them as a dependency.

So, reset MW services, SMW services and assorted static singletons
before each integration test. This conveniently allows for reenabling
several tests that had been previously disabled due to running into
exactly the kind of problems described above. Since resetting the MW
service container also resets hook handlers, add appropriate hook
registration logic to tests that previously implicitly relied on hook
registrations "leaking" over from another test (mostly Query-related
ones). As we're here, declare a bunch of missing properties in test
classes that pollute the output on PHP > 8.2.

On MW 1.39 (and possibly older), the test output appears to be buffered
by default, which is inconvenient because one cannot see the progress
until the whole suite has concluded. Set stderr="true" in the config to
rectify that (matching what core tests already do), and update the
environment reporter to write to stderr as well.

This change does not yet make the suite green on MW 1.39 - there are
various failures that need to be investigated. However, it should fix
most issues that are purely test setup-related.