Remove redundant CI configuration
The php-8 workflow file is identical to the main workflow file except
for the Ubuntu runner version and the lack of the MW patch from
9c2aa6e8215af4204f4fa18fca7ac28fd9249ac2. Since both workflows use
containers to setup MediaWiki, the PHP version used is governed by those
containers. We can therefore remove one of these configs as redundant.

Also clean up the aforementioned MW core patch now that we're here, as
it should be no longer necessary.