Fix DISPLAYTITLE sort behavior on MW 1.38+
Parser::getDefaultSort would fall back to an empty string if no
defaultsort was provided, but in newer MW we have to use a page property
with a null default value to obtain the defaultsort value.
DisplayTitlePropertyAnnotator however expects an empty string for the
"no defaultsort" case.

The expectation here is that this should make p-0416.json pass on MW
1.38 and newer.

Closes #5585