Run the unit tests again
In #5039, the unit tests were moved from tests/phpunit/Unit to directly
under tests/phpunit, but the PHPUnit configuration was not updated to
match. Due to this, 5k+ unit tests have not been running for ~3 years.

So, update the PHPUnit configuration to run these tests again and make
fixes as needed. Due to the size of the changes, where nontrivial
changes to non-test code would have been required to fix the tests, I
opted to skip the test instead with a "SUT needs refactoring" message so
that the necessary changes could be made later without complicating the
review.

Many tests emit deprecation warnings on newer PHP. I only fixed missing
property declarations in test code in this patch; the rest will need
investigation to determine whether they are the result of incomplete or
invalid stubbings or actual issues in the code.