Fix invalid nesting of queries for ElasticSearch 7.x
A condition without a type will be converted to a list, which is no longer a valid for ElasticSearch 7.x, and leads to the following error:

```
ParsingException[Failed to parse]; nested: XContentParseException[[1:335] [bool] failed to parse field [must]]; nested: IllegalStateException[expected value but got [START_ARRAY]];; org.elasticsearch.common.xcontent.XContentParseException: [1:335] [bool] failed to parse field [must]
```

By setting the type to 'must', the array is wrapped in a `bool.must` block. This is valid, and should be semantically identical.