Update PropertyTableUpdater: Fix "Data truncated for column 'o_serialized' at row 1" errors
Due to https://github.com/wikimedia/mediawiki/commit/551ec29ea676cc73cd127b2f9cfc9bedcfc9fdc5 it now throws an error. It's to mimic behaviour for when mysql is in strict mode and is switched on for CI by default. Instead of switching this off for CI let's just put this in a try and catch.

This doesn't fix the truncation problem but it does bring back previous behaviour for MW 1.42+.