DIUriHandler: Fix "Data truncated for column 'o_serialized' at row 1"  errors (#5961)
* DIUriHandler: Fix "Data truncated for column 'o_serialized' at row 1" errors

Due to https://github.com/wikimedia/mediawiki/commit/551ec29ea676cc73cd127b2f9cfc9bedcfc9fdc5 it now throws an error. It's to mimic behaviour for when mysql is in strict mode and is switched on for CI by default.

Instead of relying on MySQL to do the truncation which in strict mode throws an error, we do the truncation software side.

This makes it a noop for installs without strict mode enabled. It fixes installs that have strict mode enabled.

* Update DIUriHandler.php