5.0.0 release
Please add missing entries to the RELEASE-NOTES *outside* of this PR

Also finish up *breaking changes* that you want to be merged any time soon before the scheduled release on the 9th.

For reference, here is the 5.0.0 milestone: https://github.com/SemanticMediaWiki/SemanticMediaWiki/milestone/46 The release will not be blocked on finishing that wishlist.